// packages/openresponses-client/src/generated/types/CreateResponseBody.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { IncludeEnum } from "./IncludeEnum.ts";
import type { ItemParam } from "./ItemParam.ts";
import type { MetadataParam } from "./MetadataParam.ts";
import type { ReasoningParam } from "./ReasoningParam.ts";
import type { ResponsesToolParam } from "./ResponsesToolParam.ts";
import type { ServiceTierEnum } from "./ServiceTierEnum.ts";
import type { StreamOptionsParam } from "./StreamOptionsParam.ts";
import type { TextParam } from "./TextParam.ts";
import type { ToolChoiceParam } from "./ToolChoiceParam.ts";
import type { TruncationEnum } from "./TruncationEnum.ts";

export type CreateResponseBody = {
	model?: string | null;
	input?: (string | ItemParam[]) | null;
	previous_response_id?: string | null;
	/**
	 * @type array | undefined
	 */
	include?: IncludeEnum[];
	tools?: ResponsesToolParam[] | null;
	tool_choice?: (ToolChoiceParam & any) | null;
	metadata?: (MetadataParam & any) | null;
	text?: (TextParam & any) | null;
	temperature?: number | null;
	top_p?: number | null;
	presence_penalty?: number | null;
	frequency_penalty?: number | null;
	parallel_tool_calls?: boolean | null;
	/**
	 * @description Whether to stream response events as server-sent events.
	 * @type boolean | undefined
	 */
	stream?: boolean;
	stream_options?: (StreamOptionsParam & any) | null;
	/**
	 * @description Whether to run the request in the background and return immediately.
	 * @type boolean | undefined
	 */
	background?: boolean;
	max_output_tokens?: number | null;
	max_tool_calls?: number | null;
	reasoning?: (ReasoningParam & any) | null;
	safety_identifier?: string | null;
	prompt_cache_key?: string | null;
	truncation?: TruncationEnum & any;
	instructions?: string | null;
	/**
	 * @description Whether to store the response so it can be retrieved later.
	 * @type boolean | undefined
	 */
	store?: boolean;
	service_tier?: ServiceTierEnum & any;
	top_logprobs?: number | null;
};