// packages/openresponses-client/src/generated/types/ErrorStreamingEvent.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ErrorPayload } from "./ErrorPayload.ts";

/**
 * @description A streaming event that indicated an error was emitted.
 */
export type ErrorStreamingEvent = {
	/**
	 * @description The type of the event, always `error`.
	 * @default "error"
	 * @type string
	 */
	type: "error";
	/**
	 * @description The sequence number of the event that was emitted.
	 * @type integer
	 */
	sequence_number: number;
	error: ErrorPayload & any;
};