// packages/openresponses-client/src/generated/types/ResponseCompletedStreamingEvent.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ResponseResource } from "./ResponseResource.ts";

/**
 * @description A streaming event that indicated the response was completed.
 */
export type ResponseCompletedStreamingEvent = {
	/**
	 * @description The type of the event, always `response.completed`.
	 * @default "response.completed"
	 * @type string
	 */
	type: "response.completed";
	/**
	 * @description The sequence number of the event that was emitted.
	 * @type integer
	 */
	sequence_number: number;
	response: ResponseResource & any;
};