// packages/openresponses-client/src/generated/types/ResponseContentPartDoneStreamingEvent.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { InputFileContent } from "./InputFileContent.ts";
import type { InputImageContent } from "./InputImageContent.ts";
import type { InputTextContent } from "./InputTextContent.ts";
import type { OutputTextContent } from "./OutputTextContent.ts";
import type { ReasoningTextContent } from "./ReasoningTextContent.ts";
import type { RefusalContent } from "./RefusalContent.ts";
import type { SummaryTextContent } from "./SummaryTextContent.ts";
import type { TextContent } from "./TextContent.ts";

/**
 * @description A streaming event that indicated a content part was completed.
 */
export type ResponseContentPartDoneStreamingEvent = {
	/**
	 * @description The type of the event, always `response.content_part.done`.
	 * @default "response.content_part.done"
	 * @type string
	 */
	type: "response.content_part.done";
	/**
	 * @description The sequence number of the event that was emitted.
	 * @type integer
	 */
	sequence_number: number;
	/**
	 * @description The ID of the item that was updated.
	 * @type string
	 */
	item_id: string;
	/**
	 * @description The index of the output item that was updated.
	 * @type integer
	 */
	output_index: number;
	/**
	 * @description The index of the content part that was completed.
	 * @type integer
	 */
	content_index: number;
	/**
	 * @description A content part that makes up an input or output item.
	 */
	part:
		| InputTextContent
		| OutputTextContent
		| TextContent
		| SummaryTextContent
		| ReasoningTextContent
		| RefusalContent
		| InputImageContent
		| InputFileContent;
};