// packages/openresponses-client/src/generated/types/ResponseCreatedStreamingEvent.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ResponseResource } from "./ResponseResource.ts";

/**
 * @description A streaming event that indicated the response was created.
 */
export type ResponseCreatedStreamingEvent = {
	/**
	 * @description The type of the event, always `response.created`.
	 * @default "response.created"
	 * @type string
	 */
	type: "response.created";
	/**
	 * @description The sequence number of the event that was emitted.
	 * @type integer
	 */
	sequence_number: number;
	response: ResponseResource & any;
};