// packages/openresponses-client/src/generated/types/ResponseFailedStreamingEvent.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ResponseResource } from "./ResponseResource.ts";

/**
 * @description A streaming event that indicated the response had failed.
 */
export type ResponseFailedStreamingEvent = {
	/**
	 * @description The type of the event, always `response.failed`.
	 * @default "response.failed"
	 * @type string
	 */
	type: "response.failed";
	/**
	 * @description The sequence number of the event that was emitted.
	 * @type integer
	 */
	sequence_number: number;
	response: ResponseResource & any;
};