// packages/openresponses-client/src/generated/types/ResponseFunctionCallArgumentsDeltaStreamingEvent.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description A streaming event that indicated function call arguments were incrementally added.
 */
export type ResponseFunctionCallArgumentsDeltaStreamingEvent = {
	/**
	 * @description The type of the event, always `response.function_call_arguments.delta`.
	 * @default "response.function_call_arguments.delta"
	 * @type string
	 */
	type: "response.function_call_arguments.delta";
	/**
	 * @description The sequence number of the event that was emitted.
	 * @type integer
	 */
	sequence_number: number;
	/**
	 * @description The ID of the tool call item that was updated.
	 * @type string
	 */
	item_id: string;
	/**
	 * @description The index of the output item that was updated.
	 * @type integer
	 */
	output_index: number;
	/**
	 * @description The arguments delta that was appended.
	 * @type string
	 */
	delta: string;
	/**
	 * @description An obfuscation string that was added to pad the event payload.
	 * @type string | undefined
	 */
	obfuscation?: string;
};